requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
confluent_kafka==2.4.0
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
joblib==1.5.1
kafka-python==2.2.15
numpy==2.0.2
prometheus_client==0.22.1
pydantic==2.11.7
pydantic_core==2.33.2
requests==2.32.4
scikit-learn==1.6.1
scipy==1.13.1
sniffio==1.3.1
starlette==0.47.1
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
